fix: isolate anonymization regression test
